Move projectExists method into projects module
